requirements/prod.txt
numpy>=1.16.4
pandas>=0.25.1
scikit-learn
torch>=1.2.0
tqdm>=4.35.0
Biopython>=1.74
tensorboard
pyyml
